Handle packed GS outputs.
PipeGS is an example of a shader in which two separate outputs are
packed into a single vector.
This causes the following declarations in the bytecode:
dcl_output o3.xy
dcl_output o3.z
which would cause output3 to be declared twice in the GLSL shader.

With this change the generated GLSL for this case will be
#ifndef VtxGeoOutput3_CREATED
#define VtxGeoOutput3_CREATED
out vec4 VtxGeoOutput3;
#define Output3 VtxGeoOutput3
#endif
#ifndef VtxGeoOutput3_CREATED
#define VtxGeoOutput3_CREATED
out vec4 VtxGeoOutput3;
#define Output3 VtxGeoOutput3
#endif