icb fixes, itof, imax, immediate32_plus_relative indexing.
The immediate constant buffer can contain a mix of integer and float
values. Determining the type of one these constants is awkward. A single
vec4 can also mix integer and float types.
In this changeset I have forced NAN and INF to zero inside the immediate
constant buffer. This will allow the shader to compile. One full
solution might be to have two immediate constant buffers, one for float
and one for int. Another solution would be to have all values in the
immediate constant buffers as hex integers and use
ARB_shader_bit_encoding to convert for opcodes which operate on floating
point values.