Started looking at improving integer support.
Decode the data type of constant buffer member variables in order to
find out which ones are integer.
Uniforms are now declared using member variable type information rather
than always being floating point vec4 arrays. Only done for the non-ubo
case at the moment.
Some support for taking a compile-time-known index into the uniform +
swizzle and working out which named uniform it is.
Keep track of which data type each temp register currently is.
Inputs and outputs are now integer if they are meant to be.