Extension fallback for query-lod, texture-gather, and atomic counters.
Enable the GL_ARB_texture_query_lod, GL_ARB_texture_gather and
GL_ARB_shader_atomic_counters extensions when those features are used
with a target language which does not support it.
Made some changes to handle the fact that the textureQueryLOD extension
function was renamed to textureQueryLod in core.