MOVC and MOV now more compatible with typeless data.
For example if the data moved by MOVC was float in once case and integer
in the other, this can be managed by making the destination of the write
an integer type and an any floating point sources will need
floatBitsToInt. When a floating point operation is performed
intBitsToFloat is used to get back to a floating point type.