Fix tests\ps2\uniformFuncParam compile errors.
tests\ps2\uniformFuncParam was broken by the changes for issue #12.
Don't need to create a shadow sampler for every texture, just the ones
used with certain opcodes. Also changed the string prepended to the
shadow sampler name from hlslcc_zcmp_ to hlslcc_zcmp because if the
sampler name started with an underscore then the shadow sampler name
would include two consecutive underscores which is reserved in GL.